requirements.txt
icecream==2.1.2
matplotlib==3.5.1
nltk==3.7
pandas==1.4.2
pandas_stubs==1.2.0.57
wmd
numpy
openai
sentence_transformers
tensorboardX
spacy
albumentations==0.4.3
opencv-python
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
torchmetrics==0.6
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
transformers==4.19.2
webdataset==0.2.5
open-clip-torch==2.7.0
gradio==3.11
kornia==0.6
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
icecream
hydra-core
evaluate
rouge
-e .